HTBHF-1680 refactor remove child from children
