+++ src/web/routes/application/children-dob/remove-child-by-index.js
@@ -1,60 +1,25 @@
-const { compose, map, isNil, pickBy, toPairs, fromPairs } = require('ramda')
-
-const INDEX_FROM_KEY_REG_EXP = /.*-(\d+).*/
-
-const getIndexFromKey = (key) => {
-  const groups = INDEX_FROM_KEY_REG_EXP.exec(key)
-
-  if (isNil(groups)) {
-    throw new Error(`Could not find index for key ${key}`)
-  }
-
-  return parseInt(groups[1], 10)
-}
-
-const isKeyAfterRemoved = (key, index) => getIndexFromKey(key) > index
-
-const decrementKey = (key) => {
-  const index = getIndexFromKey(key)
-
-  if (index <= 1) {
-    throw new Error(`Unable to decrement index for key ${key}`)
-  }
-
-  return key.replace(index, index - 1)
-}
-
-const keyDoesNotContainIndex = (index) => (val, key) => getIndexFromKey(key) !== index
-
-const handleDecrement = (index) => (pair) => {
-  const key = pair[0]
-
-  if (isKeyAfterRemoved(key, index)) {
-    return [decrementKey(key), pair[1]]
-  }
-
-  return pair
-}
+const { compose, map, toPairs, fromPairs } = require('ramda')
+const { handleDecrement } = require('./decrement-keys')
+const { getChildEntries, getNonChildEntries, getEntriesWithoutIndex } = require('./selectors')
 
 const reIndexChildren = (children, index) => compose(fromPairs, map(handleDecrement(index)), toPairs)(children)
 
+/**
+ * Remove all 'childDob-x' entries from the session object (children) where x matches the index by:
+ *  - Get all the entries from the session that start with 'childDob-'
+ *  - Remove all entries which are for the given index ('childDob-x')
+ *  - Rebuild the object with the remaining children entries re-indexed plus everything else (non-'childDob-x' entries)
+ */
 const removeChildByIndex = (children, index) => {
-  const childDobPrefix = 'childDob-'
-
-  const metaEntries = pickBy((val, key) => !key.startsWith(childDobPrefix), children)
-  const childEntries = pickBy((val, key) => key.startsWith(childDobPrefix), children)
-
-  const filteredChildren = pickBy(keyDoesNotContainIndex(index), childEntries)
+  const childEntries = getChildEntries(children)
+  const filteredChildren = getEntriesWithoutIndex(childEntries, index)
 
   return {
     ...reIndexChildren(filteredChildren, index),
-    ...metaEntries
+    ...getNonChildEntries(children)
   }
 }
 
 module.exports = {
-  isKeyAfterRemoved,
-  decrementKey,
-  keyDoesNotContainIndex,
   removeChildByIndex
 }
