+++ src/web/routes/application/children-dob/remove-child-by-index.test.js
@@ -1,45 +1,7 @@
 const test = require('tape')
-const { isKeyAfterRemoved, decrementKey, keyDoesNotContainIndex, removeChildByIndex } = require('./remove-child-by-index')
+const { removeChildByIndex } = require('./remove-child-by-index')
 
-test('isKeyAfterRemoved()', (t) => {
-  t.equal(isKeyAfterRemoved('childDob-1-day', 1), false)
-  t.equal(isKeyAfterRemoved('childDob-1-day', 2), false)
-  t.equal(isKeyAfterRemoved('childDob-1-day', 3), false)
-
-  t.equal(isKeyAfterRemoved('childDob-2-day', 1), true)
-  t.equal(isKeyAfterRemoved('childDob-2-day', 2), false)
-  t.equal(isKeyAfterRemoved('childDob-2-day', 3), false)
-
-  t.equal(isKeyAfterRemoved('childDob-3-day', 1), true)
-  t.equal(isKeyAfterRemoved('childDob-3-day', 2), true)
-  t.equal(isKeyAfterRemoved('childDob-3-day', 3), false)
-
-  t.equal(isKeyAfterRemoved('childDob-11-day', 1), true)
-  t.equal(isKeyAfterRemoved('childDob-1', 1), false)
-  t.equal(isKeyAfterRemoved('childDob-5', 1), true)
-
-  t.throws(() => isKeyAfterRemoved('childDob-day', 1), /Could not find index for key childDob-day/)
-  t.throws(() => isKeyAfterRemoved('inputCount', 1), /Could not find index for key inputCount/)
-  t.end()
-})
-
-test('decrementKey()', (t) => {
-  t.equal(decrementKey('childDob-2-day'), 'childDob-1-day')
-  t.equal(decrementKey('childDob-11-day'), 'childDob-10-day')
-
-  t.throws(() => decrementKey('childDob-day'), /Could not find index for key childDob-day/)
-  t.throws(() => decrementKey('inputCount'), /Could not find index for key inputCount/)
-  t.throws(() => decrementKey('childDob-1-day'), /Unable to decrement index for key childDob-1-day/)
-  t.end()
-})
-
-test('keyDoesNotContainIndex()', (t) => {
-  t.equal(keyDoesNotContainIndex(1)('value', 'childDob-1-day'), false)
-  t.equal(keyDoesNotContainIndex(2)('value', 'childDob-1-day'), true)
-  t.end()
-})
-
-test.only('removeChildByIndex', (t) => {
+test('removeChildByIndex', (t) => {
   const children = {
     'childDob-1-day': '1',
     'childDob-1-month': '1',
