+++ src/web/routes/application/children-dob/selectors.js
@@ -1,0 +1,14 @@
+const { pickBy } = require('ramda')
+const { isNotChildEntry, isChildEntry, keyDoesNotContainIndex } = require('./predicates')
+
+const getNonChildEntries = pickBy(isNotChildEntry)
+
+const getChildEntries = pickBy(isChildEntry)
+
+const getEntriesWithoutIndex = (entries, index) => pickBy(keyDoesNotContainIndex(index), entries)
+
+module.exports = {
+  getChildEntries,
+  getNonChildEntries,
+  getEntriesWithoutIndex
+}
